Extract instruction crate (#2405)
* extract instruction crate

* remove thiserror

* make bincode optional in solana-instruction

* make serde optional in solana-instruction

* fix doc tests

* fmt

* harmonize features

* fmt Cargo.toml

* clippy

* fix syscalls.rs imports

* update digests

* update digest

* unused imports

* post-rebase fmt

* fix doc links

* use workspace lints

* fmt

* make rustc_version dep optional

* update digest

* update digest

* update import

* fmt

* fix dup import

* remove dev-context-only-utils (no longer needed)

* fmt

* remove unnecessary allow()

* fix overwriting instruction syscall stubs

* fmt

* move get_processed_sibling_instruction and get_stack_height back to solana-program to avoid breaking change

* fix path

* fix typo

* move the checked_add utility function back to solana-program

* move AccountMeta to its own file

* fix bad import

* move ProcessedSiblingInstruction to syscalls.rs

* move CompiledInstruction back to solana-program

* move ProcessedSiblingInstruction back into lib.rs

* make std optional in solana-instruction

* fix required features for frozen-abi

* update digest

* update digest

* missing import

* update digest

* don't assume std is imported in frozen-abi macros

* fix import warnings

* simplify cfg usage
